fixed my account page

+++ database/migrations/2023_10_31_151833_create_s_m_s_messages_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('s_m_s_messages', function (Blueprint $table) {
             $table->id();
             $table->integer('job_id')->unsigned();
-            $table->integer('recipient');
+            $table->bigInteger('recipient');
             $table->string('message');
             $table->string('status')->default('scheduled');
             $table->text('error')->nullable();
